+++ src/compas_timber/assembly/__init__.py
@@ -1,4 +1,0 @@
-from .assembly import TimberAssembly
-from .assembly_from_surface import SurfaceAssembly
-
-__all__ = ["TimberAssembly", "SurfaceAssembly"]