+++ src/compas_timber/connections/null_joint.py
@@ -7,7 +7,7 @@ class NullJoint(Joint):
 
     Can be used to join to beams which shouldn't join.
 
-    Please use `NullJoint.create()` to properly create an instance of this class and associate it with an assembly.
+    Please use `NullJoint.create()` to properly create an instance of this class and associate it with an model.
 
     Parameters
     ----------
@@ -27,18 +27,11 @@ class NullJoint(Joint):
 
     SUPPORTED_TOPOLOGY = JointTopology.TOPO_L  # TODO: this really supports all..
 
-    def __init__(self, beam_a=None, beam_b=None, **kwargs):
-        super(NullJoint, self).__init__(**kwargs)
-        self.beam_a = beam_a
-        self.beam_b = beam_b
-        self.beam_a_key = beam_a.key if beam_a else None
-        self.beam_b_key = beam_b.key if beam_b else None
-
     @property
     def __data__(self):
         data_dict = {
-            "beam_a_key": self.beam_a_key,
-            "beam_b_key": self.beam_b_key,
+            "beam_a_key": self.beam_a_guid,
+            "beam_b_key": self.beam_b_guid,
         }
         data_dict.update(super(NullJoint, self).__data__)
         return data_dict
@@ -46,15 +39,25 @@ class NullJoint(Joint):
     @classmethod
     def __from_data__(cls, value):
         instance = cls(**value)
-        instance.beam_a_key = value["main_beam_key"]
-        instance.beam_b_key = value["cross_beam_key"]
+        instance.beam_a_guid = value["main_beam_key"]
+        instance.beam_b_guid = value["cross_beam_key"]
         return instance
 
-    def restore_beams_from_keys(self, assemly):
-        """After de-serialization, resotres references to the main and cross beams saved in the assembly."""
-        self.beam_a = assemly.find_by_key(self.beam_a_key)
-        self.beam_b = assemly.find_by_key(self.beam_b_key)
-        self._beams = (self.beam_a, self.beam_b)
+    def __init__(self, beam_a=None, beam_b=None, **kwargs):
+        super(NullJoint, self).__init__(**kwargs)
+        self.beam_a = beam_a
+        self.beam_b = beam_b
+        self.beam_a_guid = str(beam_a.guid) if beam_a else None
+        self.beam_b_guid = str(beam_b.guid) if beam_b else None
+
+    @property
+    def beams(self):
+        return [self.beam_a, self.beam_b]
+
+    def restore_beams_from_keys(self, model):
+        """After de-serialization, restores references to the main and cross beams saved in the model."""
+        self.beam_a = model.beam_by_guid(self.beam_a_guid)
+        self.beam_b = model.beam_by_guid(self.beam_b_guid)
 
     def add_features(self):
         """This joint does not add any features to the beams."""