+++ src/compas_timber/ghpython/components/CT_Joint_Rule_Category/code.py
@@ -1,16 +1,16 @@
-from ghpythonlib.componentbase import executingcomponent as component
-from Grasshopper.Kernel.GH_RuntimeMessageLevel import Warning
-from System.Windows.Forms import ToolStripSeparator
-from System.Windows.Forms import ToolStripMenuItem
-
 import inspect
 from collections import OrderedDict
 
+from ghpythonlib.componentbase import executingcomponent as component
+from Grasshopper.Kernel.GH_RuntimeMessageLevel import Warning
+from System.Windows.Forms import ToolStripMenuItem
+from System.Windows.Forms import ToolStripSeparator
+
 from compas_timber.connections import Joint
-from compas_timber.ghpython.ghcomponent_helpers import manage_dynamic_params
-from compas_timber.ghpython.ghcomponent_helpers import get_leaf_subclasses
-from compas_timber.ghpython.ghcomponent_helpers import rename_gh_output
 from compas_timber.ghpython import CategoryRule
+from compas_timber.ghpython.ghcomponent_helpers import get_leaf_subclasses
+from compas_timber.ghpython.ghcomponent_helpers import manage_dynamic_params
+from compas_timber.ghpython.ghcomponent_helpers import rename_gh_output
 
 
 class CategoryJointRule(component):