+++ src/compas_timber/ghpython/components/CT_Model/code.py
@@ -1,21 +1,20 @@
 from compas.scene import Scene
+from compas.tolerance import TOL
 from ghpythonlib.componentbase import executingcomponent as component
 from Grasshopper.Kernel.GH_RuntimeMessageLevel import Warning
 
-from compas_timber.assembly import TimberAssembly
-from compas_timber.consumers import BrepGeometryConsumer
+from compas_timber.connections import BeamJoinningError
 from compas_timber.connections import ConnectionSolver
 from compas_timber.connections import JointTopology
-from compas_timber.connections import BeamJoinningError
-from compas_timber.connections import XHalfLapJoint
-from compas_timber.connections import TButtJoint
 from compas_timber.connections import LMiterJoint
-from compas_timber.ghpython import JointDefinition
+from compas_timber.connections import TButtJoint
+from compas_timber.connections import XHalfLapJoint
 from compas_timber.ghpython import CategoryRule
-from compas_timber.ghpython import TopologyRule
-from compas_timber.ghpython import DirectRule
 from compas_timber.ghpython import DebugInfomation
-
+from compas_timber.ghpython import DirectRule
+from compas_timber.ghpython import JointDefinition
+from compas_timber.ghpython import TopologyRule
+from compas_timber.model import TimberModel
 
 JOINT_DEFAULTS = {
     JointTopology.TOPO_X: XHalfLapJoint,
@@ -24,19 +23,7 @@ JOINT_DEFAULTS = {
 }
 
 
-class Assembly(component):
-    def __init__(self):
-        # maintains relationship of old_beam.id => new_beam_obj for referencing
-        # lets us modify copies of the beams while referencing them using their old identities.
-        self._beam_map = {}
-
-    def _get_copied_beams(self, old_beams):
-        """For the given old_beams returns their respective copies."""
-        new_beams = []
-        for beam in old_beams:
-            new_beams.append(self._beam_map[id(beam)])
-        return new_beams
-
+class ModelComponent(component):
     def get_joints_from_rules(self, beams, rules, topologies):
         if not isinstance(rules, list):
             rules = [rules]
@@ -88,7 +75,10 @@ class Assembly(component):
                         self.AddRuntimeMessage(
                             Warning,
                             msg.format(
-                                beam_a.key, beam_b.key, JointTopology.get_name(detected_topo), rule.joint_type.__name__
+                                beam_a.guid,
+                                beam_b.guid,
+                                JointTopology.get_name(detected_topo),
+                                rule.joint_type.__name__,
                             ),
                         )
                         continue
@@ -97,8 +87,8 @@ class Assembly(component):
                         self.AddRuntimeMessage(
                             Warning,
                             msg.format(
-                                beam_a.key,
-                                beam_b.key,
+                                beam_a.guid,
+                                beam_b.guid,
                                 JointTopology.get_name(detected_topo),
                                 [JointTopology.get_name(topo) for topo in rule.topos],
                             ),
@@ -129,10 +119,17 @@ class Assembly(component):
             self.AddRuntimeMessage(Warning, "Input parameter JointRules failed to collect data")
         if not (Beams):  # shows beams even if no joints are found
             return
+        if MaxDistance is None:
+            MaxDistance = TOL.ABSOLUTE  # compared to calculted distance, so shouldn't be just 0.0
 
-        Assembly = TimberAssembly()
+        Model = TimberModel()
         debug_info = DebugInfomation()
-
+        for beam in Beams:
+            # prepare beams for downstream processing
+            beam.remove_features()
+            beam.remove_blank_extension()
+            beam.debug_info = []
+            Model.add_beam(beam)
         topologies = []
         solver = ConnectionSolver()
         found_pairs = solver.find_intersecting_pairs(Beams, rtree=True, max_distance=MaxDistance)
@@ -141,16 +138,9 @@ class Assembly(component):
             detected_topo, beam_a, beam_b = solver.find_topology(beam_a, beam_b, max_distance=MaxDistance)
             if not detected_topo == JointTopology.TOPO_UNKNOWN:
                 topologies.append({"detected_topo": detected_topo, "beam_a": beam_a, "beam_b": beam_b})
-        Assembly.set_topologies(topologies)
+        Model.set_topologies(topologies)
 
-        self._beam_map = {}
-        beams = [b for b in Beams if b is not None]
-        for beam in beams:
-            c_beam = beam.copy()
-            Assembly.add_beam(c_beam)
-            self._beam_map[id(beam)] = c_beam
-        beams = Assembly.beams
-
+        beams = Model.beams
         joints = self.get_joints_from_rules(beams, JointRules, topologies)
 
         if joints:
@@ -158,12 +148,12 @@ class Assembly(component):
             joints = [j for j in joints if j is not None]
             # apply reversed. later joints in orginal list override ealier ones
             for joint in joints[::-1]:
-                beams_to_pair = self._get_copied_beams(joint.beams)
+                beams_to_pair = joint.beams
                 beam_pair_ids = set([id(beam) for beam in beams_to_pair])
                 if beam_pair_ids in handled_beams:
                     continue
                 try:
-                    joint.joint_type.create(Assembly, *beams_to_pair, **joint.kwargs)
+                    joint.joint_type.create(Model, *beams_to_pair, **joint.kwargs)
                 except BeamJoinningError as bje:
                     debug_info.add_joint_error(bje)
                 else:
@@ -172,24 +162,21 @@ class Assembly(component):
         if Features:
             features = [f for f in Features if f is not None]
             for f_def in features:
-                beams_to_modify = self._get_copied_beams(f_def.beams)
-                for beam in beams_to_modify:
+                for beam in f_def.beams:
                     beam.add_features(f_def.feature)
 
         Geometry = None
         scene = Scene()
-        if CreateGeometry:
-            vis_consumer = BrepGeometryConsumer(Assembly)
-            for result in vis_consumer.result:
-                scene.add(result.geometry)
-                if result.debug_info:
-                    debug_info.add_feature_error(result.debug_info)
-        else:
-            for beam in Assembly.beams:
+        for beam in Model.beams:
+            if CreateGeometry:
+                scene.add(beam.geometry)
+                if beam.debug_info:
+                    debug_info.add_feature_error(beam.debug_info)
+            else:
                 scene.add(beam.blank)
 
         if debug_info.has_errors:
             self.AddRuntimeMessage(Warning, "Error found during joint creation. See DebugInfo output for details.")
 
         Geometry = scene.draw()
-        return Assembly, Geometry, debug_info
+        return Model, Geometry, debug_info